Formatting

+++ src/com/mtit/service/ServicePublish.java
@@ -2,6 +2,6 @@ package com.mtit.service;
 
 public interface ServicePublish {
 	
-	public String publishService(String type, int distance);
+	public String publishService(String type, double distance);
 
 }
