+++ src/com/mtit/service/ServicePublishImpl.java
@@ -3,7 +3,7 @@ package com.mtit.service;
 public class ServicePublishImpl implements ServicePublish {
 
 	@Override
-	public String publishService(String type, int distance) {
+	public String publishService(String type, double distance) {
 		AmbulanceModel ambulanceModel = new AmbulanceModel(type);
 		AmbulanceService ambulanceService = new AmbulanceService();
